chore: 正しく disconnect されないので修正
