+++ pkg/api/prisma/scripts/prod/0000-seed.ts
@@ -1,16 +1,6 @@
-import { PrismaClient } from '@prisma/client'
+import type { PrismaClient } from '@prisma/client'
 
-const prisma = new PrismaClient()
-export const main = async () => {
+export const main = async (_prisma: PrismaClient) => {
   console.warn('TODO: do something')
   // ...
-}
-
-if (require.main === module) {
-  main()
-    .catch((e) => {
-      console.error(e)
-      process.exit(1)
-    })
-    .finally(() => prisma.$disconnect())
 }
