+++ pkg/api/prisma/scripts/seed/specials.ts
@@ -1,9 +1,8 @@
-import { PrismaClient } from '@prisma/client'
+import type {PrismaClient} from '@prisma/client'
 import { deskData, projectData, revisionData, workData } from './specials-data/basic'
 import { messageData, replyData } from './specials-data/stream'
 
-const prisma = new PrismaClient()
-export const main = async () => {
+export const main = async (prisma: PrismaClient) => {
   await Promise.all(
     projectData.map((p) =>
       prisma.project.upsert({
@@ -83,13 +82,4 @@ export const main = async () => {
       })
     )
   )
-}
-
-if (require.main === module) {
-  main()
-    .catch((e) => {
-      console.error(e)
-      process.exit(1)
-    })
-    .finally(() => prisma.$disconnect())
 }
